+++ include/EST_Chunk.h
@@ -119,7 +119,7 @@ class EST_Chunk  {
     ~EST_Chunk();
 
     EST_Chunk *operator & ();
-    void *operator new (size_t size, int bytes);
+    void *operator new (size_t size, size_t bytes);
     void operator delete (void *it);
     
     void operator ++ ()
@@ -131,9 +131,9 @@ class EST_Chunk  {
   public:
     friend class EST_ChunkPtr;
 
-    friend EST_ChunkPtr chunk_allocate(int bytes);
-    friend EST_ChunkPtr chunk_allocate(int bytes, const char *initial, int initial_len);
-    friend EST_ChunkPtr chunk_allocate(int bytes, const EST_ChunkPtr &initial, int initial_start, int initial_len);
+    friend EST_ChunkPtr chunk_allocate(size_t bytes);
+    friend EST_ChunkPtr chunk_allocate(size_t bytes, const char *initial, size_t initial_len);
+    friend EST_ChunkPtr chunk_allocate(size_t bytes, const EST_ChunkPtr &initial, size_t initial_start, size_t initial_len);
 
     friend void cp_make_updatable(EST_ChunkPtr &shared, EST_chunk_size inuse);
     friend void cp_make_updatable(EST_ChunkPtr &shared);
@@ -219,8 +219,8 @@ class EST_ChunkPtr {
     });
 
 
-    char operator [] (int i) const { return ptr->memory[i]; };
-    char &operator () (int i) CII({ 
+    char operator [] (size_t i) const { return ptr->memory[i]; };
+    char &operator () (size_t i) CII({ 
       if (ptr->count>1) 
 	{
 	  CHUNK_WARN("getting writable version of shared chunk\n");
@@ -230,9 +230,9 @@ class EST_ChunkPtr {
     });
 
     // Creating a new one
-    friend EST_ChunkPtr chunk_allocate(int size);
-    friend EST_ChunkPtr chunk_allocate(int bytes, const char *initial, int initial_len);
-    friend EST_ChunkPtr chunk_allocate(int bytes, const EST_ChunkPtr &initial, int initial_start, int initial_len);
+    friend EST_ChunkPtr chunk_allocate(size_t size);
+    friend EST_ChunkPtr chunk_allocate(size_t bytes, const char *initial, size_t initial_len);
+    friend EST_ChunkPtr chunk_allocate(size_t bytes, const EST_ChunkPtr &initial, size_t initial_start, size_t initial_len);
 
     // Make sure the memory isn`t shared.
     friend void cp_make_updatable(EST_ChunkPtr &shared, EST_Chunk::EST_chunk_size inuse);
